## Summary

Splits massive `GameService` into focused services with orchestration layer, migrates `Game.id` from `int` to `Long`, converts response DTOs to records, and introduces a `GameNotFoundException` + 404 handler. IIIF URL construction is a re-usable utility used by both card and artwork mappers.

## Changes

**Service layer split (back-end)**

- New `StartGameService` - orchestrates `GameService` -> `ArtworkService` -> `CardService` for `POST /game/start`
- `SubmitService` renamed to `BoardSubmitService`; orchestrates the full submit flow (cards -> maxScore -> hint -> artwork usage -> mark game ready)
- `GameService` slimmed to its real responsibility: the `Game` aggregate (`createGame`, `getGame`, `markReady`, `updateMaxScore`, `getGameList`)
- `recordArtworkUsage(Long gameId)` moved from `GameService` to `ArtworkService`

**Game.id int -> Long migration**

- Entity, repositories, controllers (`@PathVariable Long`), services, and all `gameId` DTO fields switched to `Long`
- Switched because the database column is `BIGINT` and Spring Data/JPA returns a Long from auto-generated IDs

**DTO -> record conversion**

- `CardResponse`, `GameResponse`, `HintResponse` converted to records
- Mappers updated to use the record-type constructor
- `@JsonProperty("spymasterPick")` on `CardResponse` to keep the front-end contract (Jackson strips `is` from Java getters but not from record components)

**Exception handler**

- New `GameNotFoundException` replaces raw `RuntimeException` throws in `GameService` and `HintService` - re-usable

**IIIF URL**

- New `IiifUrlBuilder.forArtwork(UUID)` utility since its being used by 2 mappers now

**Documentation**

- `BACK-END.md` updated
- `docs/classdiagram_back-end_concept.mmd` added to reflect the new service split, types, and naming - will be canon back-end diagram in future

Related work items: #171724

+++ src/main/java/nl/vtek/names/game/controller/HintController.java
@@ -22,7 +22,7 @@ public class HintController {
     }
 
     @GetMapping("/{gameId}")
-    public ResponseEntity<HintResponse> getHintByGameId(@PathVariable int gameId) {
+    public ResponseEntity<HintResponse> getHintByGameId(@PathVariable Long gameId) {
         return hintService.getHintByGameId(gameId)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
